requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.6.0
SQLAlchemy==2.0.35
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
bcrypt==4.0.1
email-validator==2.2.0
slowapi==0.1.8
loguru==0.7.2
pytest==8.3.2
httpx==0.27.2
